epub-maker: extract cover draft state hook from useEpubMaker
